moved schmidt_decomposition to prevent circular import
makes much more sense like this
